Release v2.0.1.43 - Bug fixes, dead code cleanup, build system fixes
Critical bugs fixed:
- Fix thread-unsafe cache pointer escape in SegmentationEngine (return copy under lock)
- Fix class map visualization encoding (hardcoded *8 -> RegionClass::Count-1)
- Fix validation range mismatches in NukeXInstance (align to parameter declarations)
- Guard ONNX-dependent code in NukeXStackInstance with #ifdef NUKEX_USE_ONNX
- Fix determinism violation: replace nth_element with std::sort in SelectBestValue fallback
- Fix float/double mismatch in PixelSelector::ComputeConsensus
- Fix PSFSubtractor::SubtractStars N full image copies (now in-place)
- Remove undeclared PSFModel::EstimateFromStars declaration

Dead code removed:
- Remove dead NXSMinClassConfidence parameter (all 7 files)
- Remove unused LayerStack/LayerDecomposer (~937 lines)
- Remove SelectionRulesEngine::FromJSON() stub
- Fix region count magic number (hardcoded 9 -> RegionClass::Count)
- Delete divergent training_data/scripts/model.py

Build system fixes:
- CMakeLists.txt: add missing sources, fix version, add __PCL_BUILDING_MODULE
- Makefile: add -fvisibility=hidden, -D_REENTRANT, standard PCL flags

+++ src/engine/PixelSelector.cpp
@@ -675,7 +675,7 @@ PixelSelector::ConsensusResult PixelSelector::ComputeConsensus(
    }
 
    // Find consensus class (highest weighted vote)
-   float maxVotes = 0.0f;
+   double maxVotes = 0.0;
    int consensusIdx = 0;
    for ( int c = 0; c < numClasses; ++c )
    {
